requirements.txt
fastapi
uvicorn
pymupdf
requests
python-dotenv
pydantic
opencv-python
numpy
pdf2image
pytesseract